Move tests to nocover
