+++ hypothesis-python/tests/cover/test_type_lookup.py
@@ -26,7 +26,6 @@ from hypothesis.errors import (
     InvalidArgument,
     ResolutionFailed,
 )
-from hypothesis.internal.compat import PYPY
 from hypothesis.internal.reflection import get_pretty_function_description
 from hypothesis.strategies._internal import types
 from hypothesis.strategies._internal.core import _from_type
@@ -405,137 +404,3 @@ def test_abstract_resolver_fallback():
 
     # which in turn means we resolve to the concrete subtype.
     assert isinstance(gen, ConcreteBar)
-
-
-A = TypeVar("A")
-B = TypeVar("B")
-C = TypeVar("C")
-D = TypeVar("D")
-
-
-class _FirstBase(Generic[A, B]):
-    pass
-
-
-class _SecondBase(Generic[C, D]):
-    pass
-
-
-# To be tested:
-
-
-class TwoGenericBases1(_FirstBase[A, B], _SecondBase[C, D]):
-    pass
-
-
-class TwoGenericBases2(_FirstBase[C, D], _SecondBase[A, B]):
-    pass
-
-
-class OneGenericOneConrete1(_FirstBase[int, str], _SecondBase[A, B]):
-    pass
-
-
-class OneGenericOneConrete2(_FirstBase[A, B], _SecondBase[float, bool]):
-    pass
-
-
-class MixedGenerics1(_FirstBase[int, B], _SecondBase[C, bool]):
-    pass
-
-
-class MixedGenerics2(_FirstBase[A, str], _SecondBase[float, D]):
-    pass
-
-
-class AllConcrete(_FirstBase[int, str], _SecondBase[float, bool]):
-    pass
-
-
-_generic_test_types = (
-    TwoGenericBases1,
-    TwoGenericBases2,
-    OneGenericOneConrete1,
-    OneGenericOneConrete2,
-    MixedGenerics1,
-    MixedGenerics2,
-    AllConcrete,
-)
-
-
-@pytest.mark.parametrize("type_", _generic_test_types)
-def test_several_generic_bases(type_):
-    with temp_registered(_FirstBase, st.builds(type_)):
-        find_any(st.builds(_FirstBase))
-
-    with temp_registered(_SecondBase, st.builds(type_)):
-        find_any(st.builds(_SecondBase))
-
-
-def var_generic_func1(obj: _FirstBase[A, B]):
-    pass
-
-
-def var_generic_func2(obj: _SecondBase[A, B]):
-    pass
-
-
-def concrete_generic_func1(obj: _FirstBase[int, str]):
-    pass
-
-
-def concrete_generic_func2(obj: _SecondBase[float, bool]):
-    pass
-
-
-def mixed_generic_func1(obj: _FirstBase[A, str]):
-    pass
-
-
-def mixed_generic_func2(obj: _SecondBase[float, D]):
-    pass
-
-
-@pytest.mark.parametrize("type_", _generic_test_types)
-@pytest.mark.parametrize(
-    "func",
-    [
-        var_generic_func1,
-        var_generic_func2,
-        concrete_generic_func1,
-        concrete_generic_func2,
-        mixed_generic_func1,
-        mixed_generic_func2,
-    ],
-)
-def test_several_generic_bases_functions(type_, func):
-    with temp_registered(_FirstBase, st.builds(type_)), temp_registered(
-        _SecondBase, st.builds(type_)
-    ):
-        find_any(st.builds(func))
-
-    with temp_registered(type_, st.builds(type_)):
-        find_any(st.builds(func))
-
-
-def wrong_generic_func1(obj: _FirstBase[A, None]):
-    pass
-
-
-def wrong_generic_func2(obj: _SecondBase[None, bool]):
-    pass
-
-
-@_skip_callables_mark
-@pytest.mark.skipif(PYPY, reason="not supported")
-@pytest.mark.parametrize(
-    "func",
-    [
-        wrong_generic_func1,
-        wrong_generic_func2,
-    ],
-)
-def test_several_generic_bases_wrong_functions(func):
-    with temp_registered(AllConcrete, st.builds(AllConcrete)):
-        with pytest.raises(ResolutionFailed):
-            st.builds(func).example()
