+++ hypothesis-python/tests/nocover/test_drypython_returns.py
@@ -13,10 +13,16 @@
 #
 # END HEADER
 
+import sys
 from typing import Generic, TypeVar
 
-from hypothesis import given, strategies as st
+import pytest
 
+from hypothesis import given, strategies as st
+from hypothesis.errors import ResolutionFailed
+from hypothesis.internal.compat import PYPY
+
+from tests.common.debug import find_any
 from tests.common.utils import temp_registered
 
 # Primitives:
@@ -28,9 +34,7 @@ _FirstType = TypeVar("_FirstType")
 _LawType = TypeVar("_LawType")
 
 
-class KindN(
-    Generic[_InstanceType, _TypeArgType1],
-):
+class KindN(Generic[_InstanceType, _TypeArgType1]):
     pass
 
 
@@ -38,11 +42,7 @@ class Lawful(Generic[_LawType]):
     """This type defines law-related operations."""
 
 
-class MappableN(
-    Generic[_FirstType],
-    # NOTE: Here's the problematic part for issue-3060:
-    Lawful["MappableN[_FirstType]"],
-):
+class MappableN(Generic[_FirstType], Lawful["MappableN[_FirstType]"]):
     """Behaves like a functor."""
 
 
@@ -52,10 +52,7 @@ class MappableN(
 _ValueType = TypeVar("_ValueType")
 
 
-class MyFunctor(
-    KindN["MyFunctor", _ValueType],
-    MappableN[_ValueType],
-):
+class MyFunctor(KindN["MyFunctor", _ValueType], MappableN[_ValueType]):
     def __init__(self, inner_value: _ValueType) -> None:
         self.inner_value = inner_value
 
@@ -64,9 +61,7 @@ class MyFunctor(
 # =============
 
 
-def target_func(
-    mappable: "MappableN[_FirstType]",
-) -> bool:
+def target_func(mappable: "MappableN[_FirstType]") -> bool:
     return isinstance(mappable, MappableN)
 
 
@@ -75,7 +70,7 @@ def test_my_mappable(source: st.DataObject) -> None:
     """
     Checks that complex types with multiple inheritance levels and strings are fine.
 
-    Regression to https://github.com/HypothesisWorks/hypothesis/issues/3060
+    Regression test for https://github.com/HypothesisWorks/hypothesis/issues/3060
     """
     # In `returns` we register all types in `__mro__`
     # to be this exact type at the moment. But here, we only need `Mappable`.
@@ -86,3 +81,130 @@ def test_my_mappable(source: st.DataObject) -> None:
         st.builds(MyFunctor),
     ):
         assert source.draw(st.builds(target_func)) is True
+
+
+A = TypeVar("A")
+B = TypeVar("B")
+C = TypeVar("C")
+D = TypeVar("D")
+
+
+class _FirstBase(Generic[A, B]):
+    pass
+
+
+class _SecondBase(Generic[C, D]):
+    pass
+
+
+# To be tested:
+
+
+class TwoGenericBases1(_FirstBase[A, B], _SecondBase[C, D]):
+    pass
+
+
+class TwoGenericBases2(_FirstBase[C, D], _SecondBase[A, B]):
+    pass
+
+
+class OneGenericOneConrete1(_FirstBase[int, str], _SecondBase[A, B]):
+    pass
+
+
+class OneGenericOneConrete2(_FirstBase[A, B], _SecondBase[float, bool]):
+    pass
+
+
+class MixedGenerics1(_FirstBase[int, B], _SecondBase[C, bool]):
+    pass
+
+
+class MixedGenerics2(_FirstBase[A, str], _SecondBase[float, D]):
+    pass
+
+
+class AllConcrete(_FirstBase[int, str], _SecondBase[float, bool]):
+    pass
+
+
+_generic_test_types = (
+    TwoGenericBases1,
+    TwoGenericBases2,
+    OneGenericOneConrete1,
+    OneGenericOneConrete2,
+    MixedGenerics1,
+    MixedGenerics2,
+    AllConcrete,
+)
+
+
+@pytest.mark.parametrize("type_", _generic_test_types)
+def test_several_generic_bases(type_):
+    with temp_registered(_FirstBase, st.builds(type_)):
+        find_any(st.builds(_FirstBase))
+
+    with temp_registered(_SecondBase, st.builds(type_)):
+        find_any(st.builds(_SecondBase))
+
+
+def var_generic_func1(obj: _FirstBase[A, B]):
+    pass
+
+
+def var_generic_func2(obj: _SecondBase[A, B]):
+    pass
+
+
+def concrete_generic_func1(obj: _FirstBase[int, str]):
+    pass
+
+
+def concrete_generic_func2(obj: _SecondBase[float, bool]):
+    pass
+
+
+def mixed_generic_func1(obj: _FirstBase[A, str]):
+    pass
+
+
+def mixed_generic_func2(obj: _SecondBase[float, D]):
+    pass
+
+
+@pytest.mark.parametrize("type_", _generic_test_types)
+@pytest.mark.parametrize(
+    "func",
+    [
+        var_generic_func1,
+        var_generic_func2,
+        concrete_generic_func1,
+        concrete_generic_func2,
+        mixed_generic_func1,
+        mixed_generic_func2,
+    ],
+)
+def test_several_generic_bases_functions(type_, func):
+    with temp_registered(_FirstBase, st.builds(type_)), temp_registered(
+        _SecondBase, st.builds(type_)
+    ):
+        find_any(st.builds(func))
+
+    with temp_registered(type_, st.builds(type_)):
+        find_any(st.builds(func))
+
+
+def wrong_generic_func1(obj: _FirstBase[A, None]):
+    pass
+
+
+def wrong_generic_func2(obj: _SecondBase[None, bool]):
+    pass
+
+
+@pytest.mark.skipif(PYPY or sys.version_info[:2] == (3, 6), reason="not supported")
+@pytest.mark.parametrize("func", [wrong_generic_func1, wrong_generic_func2])
+def test_several_generic_bases_wrong_functions(func):
+    with temp_registered(AllConcrete, st.builds(AllConcrete)):
+        with pytest.raises(ResolutionFailed):
+            st.builds(func).example()
